Major update to core functionality: replace transports filter with pre_- & http_response
With WP 4.6, the filters previously used to hook into the remote_get functions and cache are now unusable. Functionality has been moved to utilize the `pre_http_response` and `http_response` filters instead, in each case verifying that the remote call needs to be made based on the `is_cacheable_call` function.